% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno_deg_chrom.R
\name{anno_deg_chrom}
\alias{anno_deg_chrom}
\title{Annotate differentially expressed genes (DEGs) with chromosome positions}
\usage{
anno_deg_chrom(
  deg_file,
  gff_file,
  format = "auto",
  id_col = "GeneID",
  fc_col = "log2FoldChange",
  use_strand = FALSE,
  drop_unmapped = TRUE
)
}
\arguments{
\item{deg_file}{DEG table from \bold{\emph{DESeq2}} analysis.}

\item{gff_file}{Genomic structural annotation \bold{\file{GFF3/GTF}} file path.}

\item{format}{Format of GFF3/GTF file. (\bold{\emph{"auto"}}, "gff3", "gtf").}

\item{id_col}{Gene IDs column name. (\bold{\emph{"GeneID"}}).}

\item{fc_col}{Log2(fold change) column name. (\bold{\emph{"log2FoldChange"}}).}

\item{use_strand}{Whether to resput actual strand instead of \code{"*"}. (\bold{\emph{FALSE}}).}

\item{drop_unmapped}{Whether to drop DEG entries not found in annotation. (\bold{\emph{TRUE}}).}
}
\value{
A data.frame with columns: (\bold{\emph{chrom, start, end, GeneID, log2FoldChange, strand}}).
}
\description{
Annotate \bold{\emph{differentially expressed genes (DEGs)}} with chromosome positions.
}
\examples{
# Example DEGs and GFF in GAnnoViz
deg_file <- system.file(
  "extdata",
  "example.deg",
  package = "GAnnoViz")

gff_file <- system.file(
  "extdata",
  "example.gff3.gz",
  package = "GAnnoViz")

# Annotate DEGs with chromosome positions
res <- anno_deg_chrom(
  deg_file = deg_file,
  gff_file = gff_file,
  format = "auto",
  id_col = "GeneID",
  fc_col = "log2FoldChange",
  use_strand = FALSE,
  drop_unmapped = TRUE
)
head(res)

}
