% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno_fst_dmr.R
\name{anno_fst_dmr}
\alias{anno_fst_dmr}
\title{Annotate FST/DMR slide windows with genomic features}
\usage{
anno_fst_dmr(
  gff_file,
  format = "auto",
  genomic_ranges,
  chrom_col = "CHROM",
  start_col = "BIN_START",
  end_col = "BIN_END",
  upstream = 2000,
  downstream = 200,
  ignore_strand = TRUE,
  features = c("promoter", "UTR5", "gene", "exon", "intron", "CDS", "UTR3", "intergenic")
)
}
\arguments{
\item{gff_file}{Genomic structural annotation \bold{\file{GFF3/GTF}} file path.}

\item{format}{Format of GFF3/GTF file. (\bold{\emph{"auto"}}, "gff3", "gtf").}

\item{genomic_ranges}{Genomic ranges file (e.g., FST, DMR). (\bold{\emph{chromosome (prefix: chr), start, end}}).}

\item{chrom_col}{Chromosome column name. (\bold{\emph{FST: "CHROM", DMR: "chr"}}).}

\item{start_col}{Start column name. (\bold{\emph{FST: "BIN_START", DMR: "start"}}).}

\item{end_col}{End column name. (\bold{\emph{FST: "BIN_END", DMR: "end"}}).}

\item{upstream}{Promoter upstream (bp). (\bold{\emph{2000}}).}

\item{downstream}{Promoter downstream (bp). (\bold{\emph{200}}).}

\item{ignore_strand}{Whether to ignore strand when computing overlaps. (\bold{\emph{TRUE}}).}

\item{features}{Which feature types to annotate. c(\bold{\emph{"promoter", "UTR5", "gene", "exon", "intron", "CDS", "UTR3", "intergenic"}}).}
}
\value{
Data.frame with input intervals and \bold{\emph{annotation results}}.
}
\description{
Annotate \bold{\emph{FST/DMR}} slide windows with \bold{\emph{genomic features}}.
}
\examples{
# Example GFF3 file in GAnnoViz
gff_file <- system.file(
  "extdata",
  "example.gff",
  package = "GAnnoViz")

# Annotate FST
fst_table <- system.file(
    "extdata",
    "example.fst",
    package = "GAnnoViz")

res <- anno_fst_dmr(
  gff_file = gff_file,
  format = "auto",
  genomic_ranges = fst_table,
  chrom_col = "CHROM",
  start_col = "BIN_START",
  end_col = "BIN_END",
  upstream = 2000,
  downstream = 200,
  ignore_strand = TRUE,
  features = c("promoter", "UTR5", "gene", "exon", "intron", "CDS", "UTR3", "intergenic")
)
head(res)

# Annotate DMR
dmr_table <- system.file(
    "extdata",
    "example.dmr",
    package = "GAnnoViz")

res <- anno_fst_dmr(
  gff_file = gff_file,
  format = "auto",
  genomic_ranges = dmr_table,
  chrom_col = "chr",
  start_col = "start",
  end_col = "end",
  upstream = 2000,
  downstream = 200,
  ignore_strand = TRUE,
  features = c("promoter", "UTR5", "gene", "exon", "intron", "CDS", "UTR3", "intergenic")
)
head(res)

}
\author{
benben-miao
}
