% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_exons.R
\name{extract_exons}
\alias{extract_exons}
\title{Extract Exons ranges from GFF or GTF}
\usage{
extract_exons(gff_file, format = "auto", exon_info = "all")
}
\arguments{
\item{gff_file}{Genomic structural annotation \bold{\file{GFF3/GTF}} file path.}

\item{format}{Format of GFF3/GTF file. (\bold{\emph{"auto"}}, "gff3", "gtf").}

\item{exon_info}{Exon information. (\bold{\emph{"all"}}, "chrom_id", "exon_range").}
}
\value{
A \bold{\emph{GRanges}} object.
}
\description{
Extract \bold{\emph{Exons}} ranges from GFF or GTF.
}
\examples{
# Example GFF3 file in GAnnoViz
gff_file <- system.file(
  "extdata",
  "example.gff3.gz",
  package = "GAnnoViz")

# Extract Exons
exons <- extract_exons(
  gff_file = gff_file,
  format = "auto",
  exon_info = "all")
exons

# Exon info: exon_range
exon_range <- extract_exons(
  gff_file = gff_file,
  format = "auto",
  exon_info = "exon_range")
head(exon_range)

}
\author{
benben-miao
}
