% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_genes.R
\name{extract_genes}
\alias{extract_genes}
\title{Extract genes information from genomic structural annotation GFF or GTF}
\usage{
extract_genes(gff_file, format = "auto", gene_info = "all")
}
\arguments{
\item{gff_file}{Genomic structural annotation GFF3/GTF file path.}

\item{format}{Format of GFF3/GTF file. (\emph{"auto"}, "gff3", "gtf").}

\item{gene_info}{Gene information. (\emph{"all"}, "chrom_id", "gene_id", "gene_range").}
}
\value{
A \emph{GRanges} object.
}
\description{
Extract genes information from genomic structural annotation GFF or GTF.
}
\examples{
# Example GFF3 file in SlideAnno
gff_file <- system.file(
  "extdata",
  "example.gff",
  package = "SlideAnno")

# Extract Genes
genes <- extract_genes(
  gff_file = gff_file,
  format = "auto",
  gene_info = "all")
genes

# Gene info: gene_id
gene_id <- extract_genes(
  gff_file = gff_file,
  format = "auto",
  gene_info = "gene_id")
head(gene_id)

# Gene info: gene_range
gene_range <- extract_genes(
  gff_file = gff_file,
  format = "auto",
  gene_info = "gene_range")
head(gene_range)

}
\author{
benben-miao
}
