% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_mrnas.R
\name{extract_mrnas}
\alias{extract_mrnas}
\title{Extract mRNA transcripts from GFF or GTF}
\usage{
extract_mrnas(gff_file, format = "auto", mrna_info = "all")
}
\arguments{
\item{gff_file}{Genomic structural annotation \bold{\file{GFF3/GTF}} file path.}

\item{format}{Format of GFF3/GTF file. (\bold{\emph{"auto"}}, "gff3", "gtf").}

\item{mrna_info}{mRNA information. (\bold{\emph{"all"}}, "chrom_id", "mrna_id", "mrna_range").}
}
\value{
A \bold{\emph{GRanges}} object.
}
\description{
Extract \bold{\emph{mRNA}} transcripts from GFF or GTF.
}
\examples{
# Example GFF3 file in GAnnoViz
gff_file <- system.file(
  "extdata",
  "example.gff3.gz",
  package = "GAnnoViz")

# Extract mRNAs
mrnas <- extract_mrnas(
  gff_file = gff_file,
  format = "auto",
  mrna_info = "all")
mrnas

# mRNA info: mrna_range
mrna_range <- extract_mrnas(
  gff_file = gff_file,
  format = "auto",
  mrna_info = "mrna_range")
head(mrna_range)

}
\author{
benben-miao
}
