% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_utr3.R
\name{extract_utr3}
\alias{extract_utr3}
\title{Extract 3'UTR ranges from GFF or GTF}
\usage{
extract_utr3(gff_file, format = "auto", utr3_info = "all")
}
\arguments{
\item{gff_file}{Genomic structural annotation \bold{\file{GFF3/GTF}} file path.}

\item{format}{Format of GFF3/GTF file. (\bold{\emph{"auto"}}, "gff3", "gtf").}

\item{utr3_info}{mRNA information. (\bold{\emph{"all"}}, "chrom_id", "utr3_id", "utr3_range").}
}
\value{
A \bold{\emph{GRanges}} object.
}
\description{
Extract \bold{\emph{3'UTR}} ranges from GFF or GTF.
}
\examples{
# Example GFF3 file in GAnnoViz
gff_file <- system.file(
  "extdata",
  "example.gff3.gz",
  package = "GAnnoViz")

# Extract 3'UTR
utr3 <- extract_utr3(
  gff_file = gff_file,
  format = "auto",
  utr3_info = "all")
utr3

# 3'UTR info: utr3_range
utr3_range <- extract_utr3(
  gff_file = gff_file,
  format = "auto",
  utr3_info = "utr3_range")
head(utr3_range)

}
\author{
benben-miao
}
