% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_utr5.R
\name{extract_utr5}
\alias{extract_utr5}
\title{Extract 5'UTR ranges from GFF or GTF}
\usage{
extract_utr5(gff_file, format = "auto", utr5_info = "all")
}
\arguments{
\item{gff_file}{Genomic structural annotation \bold{\file{GFF3/GTF}} file path.}

\item{format}{Format of GFF3/GTF file. (\bold{\emph{"auto"}}, "gff3", "gtf").}

\item{utr5_info}{mRNA information. (\bold{\emph{"all"}}, "chrom_id", "utr5_id", "utr5_range").}
}
\value{
A \bold{\emph{GRanges}} object.
}
\description{
Extract \bold{\emph{5'UTR}} ranges from GFF or GTF.
}
\examples{
# Example GFF3 file in GAnnoViz
gff_file <- system.file(
  "extdata",
  "example.gff",
  package = "GAnnoViz")

# Extract 5'UTR
utr5 <- extract_utr5(
  gff_file = gff_file,
  format = "auto",
  utr5_info = "all")
utr5

# 5'UTR info: utr5_id
utr5_id <- extract_utr5(
  gff_file = gff_file,
  format = "auto",
  utr5_info = "utr5_id")
head(utr5_id)

# 5'UTR info: utr5_range
utr5_range <- extract_utr5(
  gff_file = gff_file,
  format = "auto",
  utr5_info = "utr5_range")
head(utr5_range)

}
\author{
benben-miao
}
