% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz-package.R
\docType{package}
\name{GAnnoViz-package}
\alias{GAnnoViz}
\alias{GAnnoViz-package}
\title{GAnnoViz: Genomic Annotation and Visualization}
\description{
The \pkg{GAnnoViz} package provides tools for genomic annotation using
sliding windows and GFF/GTF parsing.
}
\section{Main Functions}{

\itemize{
\item \code{\link{extract_genes}}: Extract gene annotations from GFF/GTF.
\item \code{\link{extract_exons}}: Extract exon regions.
}
}

\section{Utilities}{

\itemize{
\item \code{\link{extract_mrnas}}: Extract mRNA transcripts.
\item \code{\link{extract_cds}}: Extract coding sequences (CDS).
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/benben-miao/GAnnoViz}
  \item Report bugs at \url{https://github.com/benben-miao/GAnnoViz/issues}
}

}
\author{
\strong{Maintainer}: Benben Miao \email{benben.miao@outlook.com}

}
\keyword{internal}
