% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_chrom_deg.R
\name{plot_chrom_deg}
\alias{plot_chrom_deg}
\title{Plot differentially expressed genes (DEGs) hyper/hypo distributions by chromosome}
\usage{
plot_chrom_deg(
  deg_file,
  gff_file,
  format = "auto",
  id_col = "GeneID",
  fc_col = "log2FoldChange",
  violin_scale = "count",
  violin_border = 0.5,
  point_shape = 16,
  point_size = 2,
  jitter_width = 0.2,
  hyper_color = "#ff000088",
  hypo_color = "#00880088"
)
}
\arguments{
\item{deg_file}{DEG table from \bold{\emph{DESeq2}} analysis.}

\item{gff_file}{Genomic structural annotation \bold{\file{GFF3/GTF}} file path.}

\item{format}{Format of GFF3/GTF file. (\bold{\emph{"auto"}}, "gff3", "gtf").}

\item{id_col}{Gene IDs column name. (\bold{\emph{"GeneID"}}).}

\item{fc_col}{Log2(fold change) column name. (\bold{\emph{"log2FoldChange"}}).}

\item{violin_scale}{Violin scale mode. (\bold{\emph{"count"}}, "area", "width").}

\item{violin_border}{Violin border width. (\bold{\emph{0.5}}).}

\item{point_shape}{Points shape (0-25). (\bold{\emph{16}}).}

\item{point_size}{Point size. (\bold{\emph{2}}).}

\item{jitter_width}{Horizontal jitter width. (\bold{\emph{0.2}}).}

\item{hyper_color}{Color for up-regulated points. (\bold{\emph{"#ff000088"}}).}

\item{hypo_color}{Color for down-regulated points. (\bold{\emph{"#00880088"}}).}
}
\value{
A \bold{\emph{ggplot object}} of chromosome-wise DEG distributions.
}
\description{
Plot \bold{\emph{differentially expressed genes (DEGs)}} hyper/hypo distributions by chromosomes.
}
\examples{
# DEG results from DESeq2
deg_file <- system.file(
    "extdata",
    "example.deg",
    package = "GAnnoViz")

# Genomic structure annotation
gff_file <- system.file(
    "extdata",
    "example.gff",
    package = "GAnnoViz")

# Plot
plot_chrom_deg(
  deg_file = deg_file,
  gff_file = gff_file,
  format = "auto",
  id_col = "GeneID",
  fc_col = "log2FoldChange",
  violin_scale = "count",
  violin_border = 0.5,
  point_shape = 16,
  point_size = 2,
  jitter_width = 0.2,
  hyper_color = "#ff000088",
  hypo_color = "#00880088"
)

}
