% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_chrom_genes.R
\name{plot_chrom_genes}
\alias{plot_chrom_genes}
\title{Plot chromosome structures and gene annotation}
\usage{
plot_chrom_genes(
  gff_file,
  gene_table,
  format = "auto",
  annotate = "id",
  orientation = "vertical",
  bar_width = 0.6,
  chrom_alpha = 0.1,
  gene_width = 0.5,
  chrom_color = "#008888",
  gene_color = "#0088ff",
  telomere_color = "#ff0000",
  label_size = 3,
  connector_dx1 = 0.2,
  connector_dx2 = 0.2,
  min_gap_frac = 0.02
)
}
\arguments{
\item{gff_file}{Genomic structural annotation \bold{\file{GFF3/GTF}} file path.}

\item{gene_table}{A data.frame: first column gene ID, second column gene name.}

\item{format}{Format of GFF3/GTF file. (\bold{\emph{"auto"}}, "gff3", "gtf").}

\item{annotate}{Which to annotate on the right: (\bold{\emph{"id"}}, "name").}

\item{orientation}{Coordinate orientation: (\bold{\emph{"vertical"}}, "horizontal").}

\item{bar_width}{Chromosome bars relative width. (\bold{\emph{0.6}}).}

\item{chrom_alpha}{Chromosome bars alpha. (\bold{\emph{0.1}}).}

\item{gene_width}{Gene bar relative width. (\bold{\emph{0.5}}).}

\item{chrom_color}{Chromosome bar color. (\bold{\emph{"#008888"}}).}

\item{gene_color}{Gene rectangle color. (\bold{\emph{"#0088ff"}}).}

\item{telomere_color}{Telomere color. (\bold{\emph{"#ff0000"}}).}

\item{label_size}{Label text size. (\bold{\emph{3}}).}

\item{connector_dx1}{First connector horizontal offset. (\bold{\emph{0.15}}).}

\item{connector_dx2}{Second connector horizontal offset. (\bold{\emph{0.15}}).}

\item{min_gap_frac}{Minimum vertical gap between labels (fraction of chromosome length). (\bold{\emph{0.02}}).}
}
\value{
A ggplot object of chromosomes with gene labels.
}
\description{
Plot \bold{\emph{chromosome structures}} and gene annotation.
}
\examples{
# Example GFF3 file in GAnnoViz
gff_file <- system.file(
  "extdata",
  "example.gff",
  package = "GAnnoViz")

genes <- data.frame(
  gene_id = c("HdF029609", "HdF029610"),
  gene_name = c("GeneA", "GeneB"))

# Vertical, annotate by name
plot_chrom_genes(
  gff_file = gff_file,
  gene_table = genes,
  annotate = "name",
  orientation = "vertical")

# Horizontal, annotate by id
plot_chrom_genes(
  gff_file = gff_file,
  gene_table = genes,
  annotate = "id",
  orientation = "horizontal")

}
