% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_chrom_heatmap.R
\name{plot_chrom_heatmap}
\alias{plot_chrom_heatmap}
\title{Plot genomic feature density heatmap}
\usage{
plot_chrom_heatmap(
  gff_file,
  format = "auto",
  feature = "gene",
  bin_size = 1e+06,
  orientation = "horizontal",
  palette = c("#ffffff", "#0055aa"),
  alpha = 0.9
)
}
\arguments{
\item{gff_file}{Genomic structural annotation \bold{\file{GFF3/GTF}} file path.}

\item{format}{Format of GFF3/GTF file. (\bold{\emph{"auto"}}, "gff3", "gtf").}

\item{feature}{Genomic feature to quantify. (\bold{\emph{"gene"}}, "exon", "CDS", "promoter").}

\item{bin_size}{Window size (bp) for density calculation. (\bold{\emph{1e6}}).}

\item{orientation}{Coordinate orientation. (\bold{\emph{"horizontal"}}, "vertical").}

\item{palette}{Continuous color palette for density. (\bold{\emph{c("#ffffff", "#0055aa")}}).}

\item{alpha}{Tile alpha for density heatmap. (\bold{\emph{0.9}}).}
}
\value{
A \bold{\emph{ggplot object}} showing genomic feature density heatmap.
}
\description{
Plot genomic feature density \bold{\emph{heatmap}}.
}
\examples{
# Example GFF3 file in GAnnoViz
gff_file <- system.file(
  "extdata",
  "example.gff3.gz",
  package = "GAnnoViz")

# Gene density heatmap
plot_chrom_heatmap(
  gff_file = gff_file,
  format = "auto",
  feature = "gene",
  bin_size = 1e6,
  orientation = "horizontal",
  palette = c("#ffffff", "#0055aa"),
  alpha = 0.9
)

}
\author{
benben-miao
}
