% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_chrom_structure.R
\name{plot_chrom_structure}
\alias{plot_chrom_structure}
\title{Plot chromosome structures and gene stats}
\usage{
plot_chrom_structure(
  gff_file,
  format = "auto",
  orientation = "vertical",
  bar_width = 0.6,
  chrom_alpha = 0.1,
  gene_width = 0.5,
  chrom_color = "#008888",
  gene_color = "#0088ff",
  telomere_color = "#ff0000",
  label_size = 3
)
}
\arguments{
\item{gff_file}{Genomic structural annotation \bold{\file{GFF3/GTF}} file path.}

\item{format}{Format of GFF3/GTF file. (\bold{\emph{"auto"}}, "gff3", "gtf").}

\item{orientation}{Coordinate orientation: (\bold{\emph{"vertical"}}, "horizontal").}

\item{bar_width}{Chromosome bars relative width. (\bold{\emph{0.6}}).}

\item{chrom_alpha}{Chromosome bars alpha. (\bold{\emph{0.2}}).}

\item{gene_width}{Gene bar relative width. (\bold{\emph{0.5}}).}

\item{chrom_color}{Chromosome bar color. (\bold{\emph{"#008888"}}).}

\item{gene_color}{Gene rectangle color. (\bold{\emph{"#0088ff"}}).}

\item{telomere_color}{Telomere color. (\bold{\emph{"#ff0000"}}).}

\item{label_size}{Label text size. (\bold{\emph{3}}).}
}
\value{
Plot of chromosome structure.
}
\description{
Plot \bold{\emph{chromosome structures and gene stats}}.
}
\examples{
# Example GFF3 file in SlideAnno
gff_file <- system.file(
  "extdata",
  "example.gff",
  package = "SlideAnno")

plot_chrom_structure(
  gff_file = gff_file,
  format = "auto",
  orientation = "vertical",
  bar_width = 0.6,
  chrom_alpha = 0.1,
  gene_width = 0.5,
  chrom_color = "#008888",
  gene_color = "#0088ff",
  telomere_color = "#ff0000",
  label_size = 3
)

}
