% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_deg_exp.R
\name{plot_deg_exp}
\alias{plot_deg_exp}
\title{Plot DEGs up/down along chromosomes}
\usage{
plot_deg_exp(
  deg_file,
  gff_file,
  format = "auto",
  id_col = "GeneID",
  fc_col = "log2FoldChange",
  orientation = "horizontal",
  chrom_alpha = 0.1,
  chrom_color = "#008888",
  bar_height = 0.8,
  point_size = 2,
  point_alpha = 0.3,
  up_color = "#ff0000",
  down_color = "#008800",
  mark_style = "point",
  line_width = 0.6,
  line_height = 0.8
)
}
\arguments{
\item{deg_file}{DEG table from \bold{\emph{DESeq2}} analysis.}

\item{gff_file}{Genomic structural annotation \bold{\file{GFF3/GTF}} file path.}

\item{format}{Format of GFF3/GTF file. (\bold{\emph{"auto"}}, "gff3", "gtf").}

\item{id_col}{Gene IDs column name. (\bold{\emph{"GeneID"}}).}

\item{fc_col}{Log2(fold change) column name. (\bold{\emph{"log2FoldChange"}}).}

\item{orientation}{Coordinate orientation. (\bold{\emph{"horizontal"}}, "vertical").}

\item{chrom_alpha}{Chromosome bar alpha. (\bold{\emph{0.1}}).}

\item{chrom_color}{Chromosome bar color. (\bold{\emph{"#008888"}}).}

\item{bar_height}{Chromosome bar thickness in y units. (\bold{\emph{0.8}}).}

\item{point_size}{Point size. (\bold{\emph{2}}).}

\item{point_alpha}{Point alpha. (\bold{\emph{0.3}}).}

\item{up_color}{Color for up-regulated genes. (\bold{\emph{"#ff0000"}}).}

\item{down_color}{Color for down-regulated genes. (\bold{\emph{"#008800"}}).}

\item{mark_style}{Marker style for DEGs. (\bold{\emph{"point"}}, "line").}

\item{line_width}{Line width. (\bold{\emph{0.6}}).}

\item{line_height}{Line height relative to bar radius. (\bold{\emph{0.8}}).}
}
\value{
A \bold{\emph{ggplot object}} visualizing DEGs along chromosomes.
}
\description{
Plot \bold{\emph{DEGs up/down}} along chromosomes.
}
\examples{
# DEG results from DESeq2
deg_file <- system.file(
    "extdata",
    "example.deg",
    package = "GAnnoViz")

# Genomic structure annotation
gff_file <- system.file(
    "extdata",
    "example.gff3.gz",
    package = "GAnnoViz")

# Plot DEGs along chromosomes
plot_deg_exp(
  deg_file = deg_file,
  gff_file = gff_file,
  format = "auto",
  id_col = "GeneID",
  fc_col = "log2FoldChange",
  orientation = "horizontal",
  chrom_alpha = 0.1,
  chrom_color = "#008888",
  bar_height = 0.8,
  point_size = 2,
  point_alpha = 0.3,
  up_color = "#ff0000",
  down_color = "#008800",
  mark_style = "point",
  line_width = 0.6,
  line_height = 0.8)

}
\author{
benben-miao
}
