% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dmg_chrom.R
\name{plot_dmg_chrom}
\alias{plot_dmg_chrom}
\title{Plot differentially methylated regions (DMRs) hyper/hypo distributions by chromosome}
\usage{
plot_dmg_chrom(
  dmr_file,
  violin_scale = "count",
  violin_border = 0.5,
  point_shape = 8,
  point_size = 2,
  jitter_width = 0.2,
  hyper_color = "#ff880088",
  hypo_color = "#0088ff88"
)
}
\arguments{
\item{dmr_file}{DEG table from \bold{\emph{MethylKit}} analysis.}

\item{violin_scale}{Violin scale mode. (\bold{\emph{"count"}}, "area", "width").}

\item{violin_border}{Violin border width. (\bold{\emph{0.5}}).}

\item{point_shape}{Points shape (0-25). (\bold{\emph{8}}).}

\item{point_size}{Point size. (\bold{\emph{2}}).}

\item{jitter_width}{Horizontal jitter width. (\bold{\emph{0.2}}).}

\item{hyper_color}{Color for hyper-methylated points. (\bold{\emph{"#ff000088"}}).}

\item{hypo_color}{Color for hypo-methylated points. (\bold{\emph{"#00880088"}}).}
}
\value{
A \bold{\emph{ggplot object}} of chromosome-wise DMR distributions.
}
\description{
Plot \bold{\emph{differentially methylated regions (DMRs)}} hyper/hypo distributions by chromosome.
}
\examples{
# DMR results from Methylkit
dmr_file <- system.file(
    "extdata",
    "example.dmr",
    package = "GAnnoViz")

# Plot
plot_dmg_chrom(
  dmr_file = dmr_file,
  violin_scale = "count",
  violin_border = 0.5,
  point_shape = 8,
  point_size = 2,
  jitter_width = 0.2,
  hyper_color = "#ff880088",
  hypo_color = "#0088ff88"
)

}
