% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene_stats.R
\name{plot_gene_stats}
\alias{plot_gene_stats}
\title{Plot gene stats for chromosomes}
\usage{
plot_gene_stats(
  gff_file,
  format = "auto",
  bar_width = 0.7,
  bar_color = "#0055ff55",
  lable_size = 3
)
}
\arguments{
\item{gff_file}{Path to \bold{\file{GFF3/GTF}} file as input.}

\item{format}{Format of GFF3/GTF file. (\bold{\emph{"auto"}}, "gff3", "gtf").}

\item{bar_width}{Bar width percent. (\bold{\emph{0.7}}).}

\item{bar_color}{Bar color with name or hex code. (\bold{\emph{"#0055ff55"}}).}

\item{lable_size}{Lable text size. (\bold{\emph{3}}).}
}
\value{
A plot object of gene stats for chromosomes.
}
\description{
\bold{\emph{Plot}} gene stats for chromosomes.
}
\examples{
# Example GFF3 file in GAnnoViz
gff_file <- system.file(
    "extdata",
    "example.gff",
    package = "GAnnoViz")

# Plot gene stats
plot_gene_stats(
    gff_file = gff_file,
    format = "auto",
    bar_width = 0.7,
    bar_color = "#0055ff55",
    lable_size = 3)

}
\author{
benben-miao
}
