% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene_structure.R
\name{plot_gene_structure}
\alias{plot_gene_structure}
\title{Plot gene structure (Promoter, 3'UTR, Exon, Intron, 5'UTR)}
\usage{
plot_gene_structure(
  gff_file,
  format = "auto",
  gene_id,
  x_breaks = 10,
  upstream = 2000,
  downstream = 200,
  feature_alpha = 0.8,
  intron_width = 1,
  arrow_count = 1,
  arrow_length = 5,
  arrow_unit = "pt",
  promoter_color = "#ff8800",
  utr5_color = "#008833",
  utr3_color = "#ff0033",
  exon_color = "#0033ff",
  intron_color = "#333333"
)
}
\arguments{
\item{gff_file}{Genomic structural annotation \bold{\file{GFF3/GTF}} file path.}

\item{format}{Format of GFF3/GTF file. (\bold{\emph{"auto"}}, "gff3", "gtf").}

\item{gene_id}{Gene id same as GFF3/GTF. (\bold{\emph{necessary}}).}

\item{x_breaks}{X axis breaks number. (\bold{\emph{10}}).}

\item{upstream}{Promoter upstream (bp). (\bold{\emph{2000}}).}

\item{downstream}{Promoter downstream (bp). (\bold{\emph{200}}).}

\item{feature_alpha}{Elements alpha. (\bold{\emph{0.8}}).}

\item{intron_width}{Intron line width. (\bold{\emph{1}}).}

\item{arrow_count}{Intron arrow number bold. (\bold{\emph{1}}).}

\item{arrow_length}{Intron arrows length（pt). (\bold{\emph{1}}).}

\item{arrow_unit}{Intron arrow length unit. (\bold{\emph{"pt"}}, "mm").}

\item{promoter_color}{Promoter color. (\bold{\emph{"#ff8800"}}).}

\item{utr5_color}{5'UTR color. (\bold{\emph{"#008833"}}).}

\item{utr3_color}{3'UTR color. (\bold{\emph{"#ff0033"}}).}

\item{exon_color}{Exon color. (\bold{\emph{"#0033ff"}}).}

\item{intron_color}{Exon color. (\bold{\emph{"#333333"}}).}
}
\value{
Plot of gene structure.
}
\description{
Plot gene structure (\bold{\emph{Promoter, 3'UTR, Exon, Intron, 5'UTR}}).
}
\examples{
# Example GFF3 file in SlideAnno
gff_file <- system.file("extdata", "example.gff", package = "SlideAnno")

# Plot gene structure
plot_gene_structure(
  gff_file = gff_file,
  format = "auto",
  gene_id = "HdF029609",
  upstream = 2000,
  downstream = 200,
  feature_alpha = 0.8,
  intron_width = 1,
  x_breaks = 10,
  arrow_length = 5,
  arrow_count = 1,
  arrow_unit = "pt",
  promoter_color = "#ff8800",
  utr5_color = "#008833",
  utr3_color = "#ff0033",
  exon_color = "#0033ff",
  intron_color = "#333333"
)

}
\author{
benben-miao
}
