% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_interval_flank.R
\name{plot_interval_flank}
\alias{plot_interval_flank}
\title{Gene neighborhood architecture around a focal gene}
\usage{
plot_interval_flank(
  gff_file,
  format = "auto",
  gene_id,
  flank_upstream = 2e+05,
  flank_downstream = 2e+05,
  show_promoters = TRUE,
  upstream = 2000,
  downstream = 200,
  arrow_length = 5,
  arrow_unit = "pt",
  gene_color = "#0088ff",
  promoter_color = "#ff8800",
  label_size = 3
)
}
\arguments{
\item{gff_file}{Genomic structural annotation \bold{\file{GFF3/GTF}} file path.}

\item{format}{Format of GFF3/GTF file. (\bold{\emph{"auto"}}, "gff3", "gtf").}

\item{gene_id}{Focal gene id consistent with GFF/GTF. (\bold{\emph{necessary}}).}

\item{flank_upstream}{Upstream flank window (bp). (\bold{\emph{200000}}).}

\item{flank_downstream}{Downstream flank window (bp). (\bold{\emph{200000}}).}

\item{show_promoters}{Whether to draw a promoter track. (\bold{\emph{TRUE}}).}

\item{upstream}{Promoter upstream (bp). (\bold{\emph{2000}}).}

\item{downstream}{Promoter downstream (bp). (\bold{\emph{200}}).}

\item{arrow_length}{Length of strand arrows (pt). (\bold{\emph{5}}).}

\item{arrow_unit}{Arrow length unit. (\bold{\emph{"pt"}}, "mm").}

\item{gene_color}{Gene bar color. (\bold{\emph{"#0088ff"}}).}

\item{promoter_color}{Promoter bar color. (\bold{\emph{"#ff8800"}}).}

\item{label_size}{Label text size. (\bold{\emph{3}}).}
}
\value{
A \bold{\emph{ggplot object}} of genes within a flank window around the focal gene.
}
\description{
Plot \bold{\emph{gene neighborhood}} around a focal \bold{\emph{gene}} with distances, arrows for strand, and optional promoter track.
}
\examples{
# Example GFF3 file in GAnnoViz
gff_file <- system.file(
  "extdata",
  "example.gff3.gz",
  package = "GAnnoViz")

# Neighborhood around a focal gene on its chromosome
plot_interval_flank(
  gff_file = gff_file,
  format = "auto",
  gene_id = "HdF029609",
  flank_upstream = 200000,
  flank_downstream = 200000,
  show_promoters = TRUE,
  upstream = 2000,
  downstream = 200,
  arrow_length = 5,
  arrow_unit = "pt",
  gene_color = "#0088ff",
  promoter_color = "#ff8800",
  label_size = 3
)

}
\author{
benben-miao
}
