% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_interval_structure.R
\name{plot_interval_structure}
\alias{plot_interval_structure}
\title{Plot gene structures for a genomic interval}
\usage{
plot_interval_structure(
  gff_file,
  format = "auto",
  chrom_id,
  start,
  end,
  x_breaks = 10,
  upstream = 2000,
  downstream = 200,
  feature_alpha = 0.8,
  intron_width = 1,
  arrow_count = 1,
  arrow_length = 5,
  arrow_unit = "pt",
  promoter_color = "#ff8800",
  utr5_color = "#008833",
  utr3_color = "#ff0033",
  exon_color = "#0033ff",
  intron_color = "#333333"
)
}
\arguments{
\item{gff_file}{Genomic structural annotation \bold{\file{GFF3/GTF}} file path.}

\item{format}{Format of GFF3/GTF file. (\bold{\emph{"auto"}}, "gff3", "gtf").}

\item{chrom_id}{Chromosome id (must match the annotation).}

\item{start}{Window start coordinate (bp, genomic absolute coordinate).}

\item{end}{Window end coordinate (bp, genomic absolute coordinate).}

\item{x_breaks}{X axis breaks number. (\bold{\emph{10}}).}

\item{upstream}{Promoter upstream (bp). (\bold{\emph{2000}}).}

\item{downstream}{Promoter downstream (bp). (\bold{\emph{200}}).}

\item{feature_alpha}{Elements alpha. (\bold{\emph{0.8}}).}

\item{intron_width}{Intron line width. (\bold{\emph{1}}).}

\item{arrow_count}{Intron arrow number bold. (\bold{\emph{1}}).}

\item{arrow_length}{Intron arrows length（pt). (\bold{\emph{1}}).}

\item{arrow_unit}{Intron arrow length unit. (\bold{\emph{"pt"}}, "mm").}

\item{promoter_color}{Promoter color. (\bold{\emph{"#ff8800"}}).}

\item{utr5_color}{5'UTR color. (\bold{\emph{"#008833"}}).}

\item{utr3_color}{3'UTR color. (\bold{\emph{"#ff0033"}}).}

\item{exon_color}{Exon color. (\bold{\emph{"#0033ff"}}).}

\item{intron_color}{Intron color. (\bold{\emph{"#333333"}}).}
}
\value{
Plot of genomic interval structure.
}
\description{
Plot gene structures for a \bold{\emph{genomic interval}}.
}
\examples{
# Example GFF3 file in GAnnoViz
gff_file <- system.file(
  "extdata",
  "example.gff3.gz",
  package = "GAnnoViz")

plot_interval_structure(
  gff_file = gff_file,
  format = "auto",
  chrom_id = "chr1",
  start = 950000,
  end = 1180000,
  x_breaks = 10,
  upstream = 2000,
  downstream = 200,
  feature_alpha = 0.8,
  intron_width = 1,
  arrow_count = 1,
  arrow_length = 5,
  arrow_unit = "pt",
  promoter_color = "#ff8800",
  utr5_color = "#008833",
  utr3_color = "#ff0033",
  exon_color = "#0033ff",
  intron_color = "#333333"
)

}
