% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_snp_anno.R
\name{plot_snp_anno}
\alias{plot_snp_anno}
\title{Plot genomic FST with Top-N gene annotations}
\usage{
plot_snp_anno(
  fst_file,
  gff_file,
  format = "auto",
  chrom_id,
  top_n = 20,
  orientation = "vertical",
  smooth_span = 0.5,
  fst_color = "#0088ff",
  point_size = 1,
  point_alpha = 0.3,
  label_size = 3,
  connector_dx1 = 20000,
  connector_dx2 = 40000,
  gap_frac = 0.05
)
}
\arguments{
\item{fst_file}{FST sliding window results. (\bold{\emph{CHROM, BIN_START, BIN_END, WEIGHTED_FST, N_VARIANTS}}.}

\item{gff_file}{Path to \bold{\file{GFF3/GTF}} file as input.}

\item{format}{Format of GFF3/GTF file. (\bold{\emph{"auto"}}, "gff3", "gtf").}

\item{chrom_id}{Chromosome identifier (\bold{\emph{"chr1"}}).}

\item{top_n}{Number of top genes to annotate. (\bold{\emph{20}}).}

\item{orientation}{Coordinate orientation. (\bold{\emph{"vertical"}}, "horizontal").}

\item{smooth_span}{Span for local regression smoothing. (\bold{\emph{0.1}}).}

\item{fst_color}{Point color for FST. (\bold{\emph{"#0088ff"}}).}

\item{point_size}{Point size. (\bold{\emph{1}}).}

\item{point_alpha}{Point alpha. (\bold{\emph{0.3}}).}

\item{label_size}{Text size for gene labels. (\bold{\emph{3}}).}

\item{connector_dx1}{First connector horizontal offset (bp). (\bold{\emph{2e4}}).}

\item{connector_dx2}{Second connector horizontal offset (bp). (\bold{\emph{4e4}}).}

\item{gap_frac}{Minimum vertical gap between labels (fraction of FST range). (\bold{\emph{0.05}}).}
}
\value{
A \bold{\emph{ggplot object}} representing genomic FST with Top-N gene annotations.
}
\description{
Plot genomic FST with \bold{\emph{Top-N gene annotations}}.
}
\examples{
# Example data in GAnnoViz
fst_file <- system.file(
  "extdata",
  "example.fst",
  package = "GAnnoViz")

gff_file <- system.file(
  "extdata",
  "example.gff3.gz",
  package = "GAnnoViz")

# Chromosome FST with Top-20 gene annotations on chr11
plot_snp_anno(
  fst_file = fst_file,
  gff_file = gff_file,
  format = "auto",
  chrom_id = "chr2",
  top_n = 20,
  orientation = "vertical",
  smooth_span = 0.5,
  fst_color = "#0088ff",
  point_size = 1,
  point_alpha = 0.3,
  label_size = 3,
  connector_dx1 = 2e4,
  connector_dx2 = 4e4,
  gap_frac = 0.05
)

}
\author{
benben-miao
}
