% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_cds.R
\name{extract_cds}
\alias{extract_cds}
\title{Extract CDS ranges from GFF or GTF}
\usage{
extract_cds(gff_file, format = "auto", cds_info = "all")
}
\arguments{
\item{gff_file}{Genomic structural annotation \bold{\file{GFF3/GTF}} file path.}

\item{format}{Format of GFF3/GTF file. (\bold{\emph{"auto"}}, "gff3", "gtf").}

\item{cds_info}{CDS information. (\bold{\emph{"all"}}, "chrom_id", "cds_range").}
}
\value{
A \bold{\emph{GRanges}} object.
}
\description{
Extract \bold{\emph{CDS}} ranges from GFF or GTF.
}
\examples{
# Example GFF3 file in GAnnoViz
gff_file <- system.file(
  "extdata",
  "example.gff",
  package = "GAnnoViz")

# Extract CDS
cds <- extract_cds(
  gff_file = gff_file,
  format = "auto",
  cds_info = "all")
cds

# CDS info: cds_range
cds_range <- extract_cds(
  gff_file = gff_file,
  format = "auto",
  cds_info = "cds_range")
head(cds_range)

}
\author{
benben-miao
}
