% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_promoters.R
\name{extract_promoters}
\alias{extract_promoters}
\title{Extract promoter ranges from GFF or GTF}
\usage{
extract_promoters(
  gff_file,
  format = "auto",
  upstream = 2000,
  downstream = 200,
  promoter_info = "all"
)
}
\arguments{
\item{gff_file}{Genomic structural annotation GFF3/GTF file path.}

\item{format}{Format of GFF3/GTF file. (\emph{"auto"}, "gff3", "gtf").}

\item{upstream}{Promoter upstream (bp). (\emph{2000}).}

\item{downstream}{Promoter downstream (bp). (\emph{200}).}

\item{promoter_info}{Promoter information. (\emph{"all"}, "chrom_id", "promoter_id", "promoter_range").}
}
\value{
A \emph{GRanges} object.
}
\description{
Extract promoter ranges from GFF or GTF.
}
\examples{
# Example GFF3 file in SlideAnno
gff_file <- system.file(
  "extdata",
  "example.gff",
  package = "SlideAnno")

# Extract Promoters
promoters <- extract_promoters(
  gff_file = gff_file,
  format = "auto",
  upstream = 2000,
  downstream = 200,
  promoter_info = "all")
promoters

# Promoter info: promoter_id
promoter_id <- extract_promoters(
  gff_file = gff_file,
  format = "auto",
  upstream = 2000,
  downstream = 200,
  promoter_info = "promoter_id")
head(promoter_id)

# Promoter info: promoter_range
promoter_range <- extract_promoters(
  gff_file = gff_file,
  format = "auto",
  upstream = 2000,
  downstream = 200,
  promoter_info = "promoter_range")
head(promoter_range)

}
\author{
benben-miao
}
